moved stationary/predefined dl4tsc into their own directory
