+++ sktime/contrib/deeplearning_based/dl4tsc/encoder.py
