+++ sktime/contrib/deeplearning_based/dl4tsc/mcdcnn.py
