+++ sktime/contrib/deeplearning_based/dl4tsc/mcnn.py
